Adding case where only one child exists
In the case where there was only one child element, it was not returned as an Array with only one element, but just as the Object itself. Therefore it caused a crashing error. 

In order to check for that I added two inline checks that check wether the element at the index 0 exists. If not we can assume that there is only one, because we already checked wether it is undefined earlier.